# Perf: optimize postal_parse_address implementation