# Intro

In this PR:

  * function `mismo.lib.geo.postal_parse_address` has been rewritten using my implementation 
  * function `mismo.lib.geo.tests.test_postal_benchmark.postal_parse_address__direct_import` is the same as the above one but avoids the optional import of `postal` inside its body (@NickCrews: don't worry, I was just curious if that import and the inner function declaration generated any relevant overhead... seems it is irrelevant, will remove it before merge)
  * function `mismo.lib.geo.tests.test_postal_benchmark.postal_parse_address__initial_impl` is the original implementation by @jstammers 

In this branch I also temporary restored the top-level `benchmarks` directory to manually gather execution times.
All benchmarks have been run on the full data set of 1M rows using `duckdb 0.10.0`, `ibis-framework 9.1.0` and `python 3.12.3` on the same laptop with no other programs running (except GNOME, etc).

```text
OS: Ubuntu 21.04 x86_64 
Host: Latitude 5500 
Kernel: 5.11.0-49-generic 
CPU: Intel i7-8665U (8) @ 4.800GHz 
GPU: Intel WhiskeyLake-U GT2 [UHD Graphics 620] 
Memory: 8067MiB / 15813MiB 
```

# Results

Results obtained with manual timings are as follows.

```text
$ python -m benchmarks.parse_address
noop                                took  19.6631 seconds
python_only                         took  20.6381 seconds
postal_only                         took  99.4631 seconds
postal_parse_address                took 109.7075 seconds
postal_parse_address__direct_import took 110.1048 seconds
postal_parse_address__initial_impl  took 118.7811 seconds
```

On the other hand, pytest-benchmark reports the following.

```text
$ just bench --benchmark-sort=mean --benchmark-group-by=param:nrows -k 1m mismo/lib/geo/tests/test_postal_benchmark.py
pytest --benchmark-only --benchmark-enable --benchmark-autosave --benchmark-sort=mean --benchmark-group-by=param:nrows -k 1m mismo/lib/geo/tests/test_postal_benchmark.py
======================================================================================================== test session starts =========================================================================================================
platform linux -- Python 3.12.3, pytest-8.2.2, pluggy-1.5.0
benchmark: 4.0.0 (defaults: timer=time.perf_counter disable_gc=False min_rounds=5 min_time=0.000005 max_time=1.0 calibration_precision=10 warmup=False warmup_iterations=100000)
rootdir: /home/lorenzo/devel/mismo
configfile: pyproject.toml
plugins: cov-5.0.0, benchmark-4.0.0, anyio-4.4.0
collected 24 items / 18 deselected / 6 selected                                                                                                                                                                                      

mismo/lib/geo/tests/test_postal_benchmark.py ......                                                                                                                                                                            [100%]
Saved benchmark data in: /home/lorenzo/devel/mismo/.benchmarks/Linux-CPython-3.12-64bit/0016_efaf478c0a66e11853fcf27704b1061ccff1fdff_20240719_124717.json



---------------------------------------------------------------------------------------------- benchmark 'nrows=1000000': 6 tests ---------------------------------------------------------------------------------------------
Name (time in s)                                                     Min                Max               Mean            StdDev             Median               IQR            Outliers     OPS            Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_benchmark_parse[1m-noop]                                    14.4026 (1.0)      14.7042 (1.0)      14.5671 (1.0)      0.1196 (1.0)      14.5971 (1.0)      0.1840 (1.0)           2;0  0.0686 (1.0)           5           1
test_benchmark_parse[1m-python_only]                             15.7207 (1.09)     16.1826 (1.10)     15.9631 (1.10)     0.2066 (1.73)     15.9745 (1.09)     0.3831 (2.08)          2;0  0.0626 (0.91)          5           1
test_benchmark_parse[1m-postal_only]                             67.2535 (4.67)     69.3878 (4.72)     68.0714 (4.67)     0.8035 (6.72)     67.7990 (4.64)     0.8358 (4.54)          2;0  0.0147 (0.21)          5           1
test_benchmark_parse[1m-postal_parse_address__direct_import]     73.5724 (5.11)     74.3592 (5.06)     73.9136 (5.07)     0.3081 (2.58)     73.9762 (5.07)     0.4267 (2.32)          2;0  0.0135 (0.20)          5           1
test_benchmark_parse[1m-postal_parse_address]                    73.3805 (5.09)     75.1065 (5.11)     73.9273 (5.07)     0.7074 (5.91)     73.7897 (5.06)     0.8699 (4.73)          1;0  0.0135 (0.20)          5           1
test_benchmark_parse[1m-postal_parse_address__initial_impl]      84.0357 (5.83)     87.4314 (5.95)     85.2971 (5.86)     1.5099 (12.62)    84.3496 (5.78)     2.3757 (12.91)         1;0  0.0117 (0.17)          5           1
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

Legend:
  Outliers: 1 Standard Deviation from Mean; 1.5 IQR (InterQuartile Range) from 1st Quartile and 3rd Quartile.
  OPS: Operations Per Second, computed as 1 / Mean
=========================================================================================== 6 passed, 18 deselected in 2350.17s (0:39:10) ============================================================================================
```

Values in the _Mean_ column are remarkably lower than those reported manually (which lay way outside the StdDev or the IQR).
Isn't it quite/too strange? As far as I can tell `pytest-benchmark` is running each function 5 times, are we caching something that we should not and subsequent executions run faster than just one (as in manual timings)?
Also, in pytest benchmarks the difference between `postal_parse_address` and `postal_only` is more than `5s` which is much bigger than the one from manual timings (`>1s`).

Additionally, before running the above results I recreated my virtualenv from zero to match mismo new requirements (`python>=3.10`, `ibis-framework>=9.1.0`) and the results I get are quite different also from those I reported in PR #48 where I was `ibis-framework 9.0.0` and `duckdb 1.0.0` (not sure about which version of python though).

# Questions

1) Before going forward, @NickCrews and @jstammers could you run both benchmarks on your systems and share the results?
2) @NickCrews: in light of the new results are you still willing to make the implementation change?