@lmores 2024-07-26
> maybe were accidentally doing computation twice in the non-pytest harness?

Quite sure this is not the case (also, my timings are a bit far from being halved).

Differences may be due to the fact that in non-pytest benchmarks I was using DuckDB that writes to disk, and not in memory as in pytest-benchmark... however, this is what I get when running non-pytest benchmarks using a memtable (doesn't change much, still far from the others):

```text
noop                                took  19.4047 seconds
python_only                         took  20.3601 seconds
postal_only                         took 104.8952 seconds
postal_parse_address                took 113.1792 seconds
postal_parse_address__direct_import took 119.0952 seconds
postal_parse_address__initial_impl  took 121.0536 seconds
``` 

At first I thought that pytest-benchmark library did something as simple as `endTimestamp - startTimestamp` to measure execution time, but as briefly explained [here](https://pytest-benchmark.readthedocs.io/en/latest/calibration.html) the measurement technique they employ seems to follow a more "probabilistical" approach (something like "wake up each `X` ms and see who is running").
Could it be the root cause of the differences???